refactor: improve folder structure and domain separation
