+++ lib/domain/entities/goal.dart
