+++ lib/domain/enums/goal_operation_result.dart
