+++ lib/domain/enums/reset_type.dart
