+++ lib/presentation/pages/goal_calendar/goal_calendar_content.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:haenaedda/model/goal.dart';
-import 'package:haenaedda/theme/app_spacing.dart';
+import 'package:haenaedda/domain/entities/goal.dart';
 import 'package:haenaedda/presentation/pages/goal_calendar/calendar_day_cell.dart';
 import 'package:haenaedda/presentation/pages/goal_calendar/goal_calendar_grid.dart';
 import 'package:haenaedda/presentation/pages/goal_calendar/goal_calendar_header.dart';
 import 'package:haenaedda/presentation/pages/goal_calendar/weekday_row.dart';
-import 'package:haenaedda/presentation/widgets/section_divider.dart';
 import 'package:haenaedda/presentation/view_models/goal_view_models.dart';
 import 'package:haenaedda/presentation/view_models/record_view_model.dart';
+import 'package:haenaedda/presentation/widgets/section_divider.dart';
+import 'package:haenaedda/theme/app_spacing.dart';
 
 class GoalCalendarContent extends StatefulWidget {
   final Goal goal;
