+++ lib/presentation/pages/goal_calendar/goal_pager.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:haenaedda/model/goal.dart';
+import 'package:haenaedda/domain/entities/goal.dart';
 import 'package:haenaedda/presentation/pages/goal_calendar/goal_calendar_content.dart';
 
 class GoalPager extends StatelessWidget {
