+++ lib/presentation/pages/launcher/launcher_page.dart
@@ -3,8 +3,8 @@ import 'package:provider/provider.dart';
 
 import 'package:haenaedda/presentation/pages/goal_calendar/goal_calendar_page.dart';
 import 'package:haenaedda/presentation/pages/launcher/add_first_goal_flow.dart';
-import 'package:haenaedda/presentation/widgets/loading_indicator.dart';
 import 'package:haenaedda/presentation/view_models/goal_view_models.dart';
+import 'package:haenaedda/presentation/widgets/loading_indicator.dart';
 
 class LauncherPage extends StatelessWidget {
   const LauncherPage({super.key});
