+++ lib/presentation/pages/reorder_goals/reorder_goals_page.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 
+import 'package:haenaedda/domain/entities/goal.dart';
 import 'package:haenaedda/gen_l10n/app_localizations.dart';
-import 'package:haenaedda/model/goal.dart';
-import 'package:haenaedda/theme/buttons.dart';
 import 'package:haenaedda/presentation/handlers/discard_changes_handler.dart';
-import 'package:haenaedda/presentation/widgets/discard_aware_close_button.dart';
 import 'package:haenaedda/presentation/view_models/goal_view_models.dart';
+import 'package:haenaedda/presentation/widgets/discard_aware_close_button.dart';
+import 'package:haenaedda/theme/buttons.dart';
 
 class ReorderGoalsPage extends StatefulWidget {
   const ReorderGoalsPage({super.key});
