+++ lib/theme/colors/light_theme.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:haenaedda/theme/colors/app_colors.dart';
+
 import 'package:haenaedda/theme/app_typography.dart';
+import 'package:haenaedda/theme/colors/app_colors.dart';
 import 'package:haenaedda/theme/colors/color_scheme_light.dart';
 import 'package:haenaedda/theme/colors/surface_container_colors_light.dart';
